[NUI] Add public types to replace nested types (CA1034 of StyleCop)
Renamed public types are added outside of the class to replace public
nested types (CA1034 of StyleCop) in Tizen.NUI.Components and
Tizen.NUI.Wearable.

This patch is done for types which have not been publicly opened.

The newly added event and EventArgs names follow the C# naming rules.

To identify the nested types from other types, the class name is
prepended to the nested type names if it is required.
e.g. FlexibleView.ItemClickEventArgs -> FlexibleViewItemClickedEventArgs
