+++ src/Tizen.NUI.Components/Controls/RadioButton.cs
@@ -80,10 +80,9 @@ namespace Tizen.NUI.Components
         /// <summary>
         /// Set CheckState to true after selecting RadioButton.
         /// </summary>
-        /// <since_tizen> 6 </since_tizen>
         /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        protected override void OnSelected()
+        protected override void OnSelectedChanged()
         {
             if (!IsSelected)
             {
