+++ test/Tizen.NUI.Samples/Tizen.NUI.Samples/Samples/FlexibleViewSample.cs
@@ -52,7 +52,7 @@ namespace Tizen.NUI.Samples
         }
     }
 
-    public class ListBridge : FlexibleView.Adapter
+    public class ListBridge : FlexibleViewAdapter
     {
         private List<ListItemData> mDatas;
 
@@ -73,15 +73,15 @@ namespace Tizen.NUI.Samples
             NotifyItemRemoved(position);
         }
 
-        public override FlexibleView.ViewHolder OnCreateViewHolder(int viewType)
+        public override FlexibleViewViewHolder OnCreateViewHolder(int viewType)
         {
-            FlexibleView.ViewHolder viewHolder = new FlexibleView.ViewHolder(new ListItemView());
+            FlexibleViewViewHolder viewHolder = new FlexibleViewViewHolder(new ListItemView());
             //Console.WriteLine($"OnCreateViewHolder... viewType: {viewType} viewID: {viewHolder.ItemView.ID}");
 
             return viewHolder;
         }
 
-        public override void OnBindViewHolder(FlexibleView.ViewHolder holder, int position)
+        public override void OnBindViewHolder(FlexibleViewViewHolder holder, int position)
         {
             //Console.WriteLine($"OnBindItemView... position: {position}");
             ListItemData listItemData = mDatas[position];
@@ -102,7 +102,7 @@ namespace Tizen.NUI.Samples
                 listItemView.BackgroundColor = Color.Yellow;
         }
 
-        public override void OnDestroyViewHolder(FlexibleView.ViewHolder holder)
+        public override void OnDestroyViewHolder(FlexibleViewViewHolder holder)
         {
             //Console.WriteLine($"OnDestroyViewHolder... viewID: {holder.ItemView?.ID}");
             if (holder.ItemView != null)
@@ -118,7 +118,7 @@ namespace Tizen.NUI.Samples
 
         public override void OnFocusChange(FlexibleView flexibleView, int previousFocus, int currentFocus)
         {
-            FlexibleView.ViewHolder previousFocusView = flexibleView.FindViewHolderForAdapterPosition(previousFocus);
+            FlexibleViewViewHolder previousFocusView = flexibleView.FindViewHolderForAdapterPosition(previousFocus);
             if (previousFocusView != null)
             {
                 //Console.WriteLine($"previousFocus {previousFocus.AdapterPosition}");
@@ -130,7 +130,7 @@ namespace Tizen.NUI.Samples
                 //previousFocus.SizeHeight = 60;
                 //NotifyItemChanged(previousFocus.AdapterPosition);
             }
-            FlexibleView.ViewHolder currentFocusView = flexibleView.FindViewHolderForAdapterPosition(currentFocus);
+            FlexibleViewViewHolder currentFocusView = flexibleView.FindViewHolderForAdapterPosition(currentFocus);
             if (currentFocusView != null)
             {
                 //Console.WriteLine($"currentFocus {currentFocus.AdapterPosition}");
@@ -141,12 +141,12 @@ namespace Tizen.NUI.Samples
             }
         }
 
-        public override void OnViewAttachedToWindow(FlexibleView.ViewHolder holder)
+        public override void OnViewAttachedToWindow(FlexibleViewViewHolder holder)
         {
             //Console.WriteLine($"+Attached: {holder.AdapterPosition}");
         }
 
-        public override void OnViewDetachedFromWindow(FlexibleView.ViewHolder holder)
+        public override void OnViewDetachedFromWindow(FlexibleViewViewHolder holder)
         {
             //Console.WriteLine($" --Detached: {holder.AdapterPosition}");
         }
@@ -301,19 +301,19 @@ namespace Tizen.NUI.Samples
             {
                 if (key.KeyPressedName == "Up")
                 {
-                    flexibleView.MoveFocus(FlexibleView.LayoutManager.Direction.Up);
+                    flexibleView.MoveFocus(FlexibleViewLayoutManager.Direction.Up);
                 }
                 else if (key.KeyPressedName == "Down")
                 {
-                    flexibleView.MoveFocus(FlexibleView.LayoutManager.Direction.Down);
+                    flexibleView.MoveFocus(FlexibleViewLayoutManager.Direction.Down);
                 }
                 if (key.KeyPressedName == "Left")
                 {
-                    flexibleView.MoveFocus(FlexibleView.LayoutManager.Direction.Left);
+                    flexibleView.MoveFocus(FlexibleViewLayoutManager.Direction.Left);
                 }
                 else if (key.KeyPressedName == "Right")
                 {
-                    flexibleView.MoveFocus(FlexibleView.LayoutManager.Direction.Right);
+                    flexibleView.MoveFocus(FlexibleViewLayoutManager.Direction.Right);
                 }
                 else if (key.KeyPressedName == "0")
                 {
